replace float to qreal

+++ turtlesim/include/turtlesim/turtle.h
@@ -60,7 +60,7 @@ class Turtle
  public:
   Turtle(rclcpp::Node::SharedPtr &node_handle, std::string &real_name, const QImage& turtle_image, const QPointF& pos, float orient);
 
-  bool update(double dt, QPainter& path_painter, const QImage& path_image, float canvas_width, float canvas_height);
+  bool update(double dt, QPainter& path_painter, const QImage& path_image, qreal canvas_width, qreal canvas_height);
   void paint(QPainter &painter);
 
  private:
